update: repositories setup

+++ core/data/src/commonMain/kotlin/com/gosty/data/api/responses/TVDetailResponse.kt
@@ -109,7 +109,7 @@ data class TVEpisodeToAirResponse(
     val overview: String,
 
     @SerialName("vote_average")
-    val voteAverage: Long,
+    val voteAverage: Double,
 
     @SerialName("vote_count")
     val voteCount: Long,
